initial commit for c/py simulator
